refactor: update shared public API
