+++ src/entities/painting/ui/ArtworkMiniature.ts
@@ -1,7 +1,6 @@
-import { Component, div, img, p } from "../../../shared/ui";
+import { Component, div, img, p, IMAGES_BASE_URL } from "../../../shared";
 
 import { type Painting } from "..";
-import { IMAGES_BASE_URL } from "../../../shared/api/urls";
 
 import styles from "./ArtworkMiniature.module.css";
