+++ src/features/hints/ui/HintControls.ts
@@ -1,8 +1,7 @@
-import { Component, i, input, label } from "../../../shared/ui";
+import { Component, i, input, label, isValidSetting } from "../../../shared";
 import { HintSettings, HintSettingsData } from "../model/HintSettings";
 
 import { Observer, Publisher } from "../../../entities/state";
-import { isValidSetting } from "../../../shared/lib";
 
 import styles from "./HintControls.module.css";
