--- src/shared/lib/index.ts
@@ -1,6 +1,0 @@
-export { Draggable } from "./drag-n-drop/Draggable";
-
-export { debounceListener } from "./debounceListener";
-export { findAllInstancesOf } from "./findAllInstancesOf";
-export { isValidSetting } from "./isValidSetting";
-export { assertNonNull } from "./assertNonNull";
