--- src/shared/ui/index.ts
@@ -1,10 +1,0 @@
-export { Component, type ElementFuncProps, type Props } from "./base/Component";
-export * from "./base/tags";
-export { Button } from "./button/Button";
-export { AudioButton } from "./button/AudioButton";
-export { ButtonIcon } from "./button/ButtonIcon";
-export { Form, type TextInputParams } from "./form/Form";
-export { Header } from "./header/Header";
-export { Input } from "./input/Input";
-export { Message, MessageType } from "./message/Message";
-export { Modal } from "./modal/Modal";
